task 1/4: Extract pure top-level EVM/display and fee-computation utilities from JoinKeysignViewModel into dedicated util files (#4866)
